requirements.txt
alembic==1.16.2
amqp==5.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
appdirs==1.4.4
attrs==25.3.0
audioread==3.0.1
av==14.4.0
basswood-av==15.2.1
bcrypt==4.3.0
billiard==4.2.1
blinker==1.9.0
boto3==1.38.45
botocore==1.38.45
cachetools==5.5.2
celery==5.3.4
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudpickle==3.1.1
coloredlogs==15.0.1
contourpy==1.3.2
ctranslate2==4.6.0
cycler==0.12.1
cytoolz==1.0.1
decorator==4.4.2
Deprecated==1.2.18
distro==1.9.0
dnspython==2.7.0
dora_search==0.1.12
ecdsa==0.19.1
einops==0.8.1
email_validator==2.2.0
fastapi==0.115.12
fastapi-cli==0.0.7
faster-whisper==1.1.1
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.58.4
fsspec==2025.5.1
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.25.1
google-auth==2.40.3
google-crc32c==1.7.1
google-genai==1.21.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.33.0
humanfriendly==10.0
humanize==4.12.3
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
intervaltree==3.1.0
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
kombu==5.5.4
lameenc==1.8.1
lazy_loader==0.4
lhotse==1.30.3
librosa==0.11.0
lilcom==1.8.1
limits==5.4.0
llvmlite==0.44.0
loguru==0.7.3
Mako==1.3.10
markdown-it-py==2.2.0
MarkupSafe==2.1.1
matplotlib==3.10.3
mdurl==0.1.2
more-itertools==10.7.0
moviepy==2.2.1
mpmath==1.3.0
msgpack==1.1.1
narwhals==1.43.0
numba==0.61.2
numpy==1.26.4
omegaconf==2.3.0
onnxruntime==1.22.0
openai==1.93.0
packaging==23.2
pandas==2.3.0
passlib==1.7.4
pedalboard==0.9.17
pexels-api-with-video-searches==1.0.1
pillow==11.2.1
platformdirs==4.3.8
pooch==1.8.2
proglog==0.1.12
progressbar==2.5
prompt_toolkit==3.0.51
proto-plus==1.26.1
protobuf==6.31.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
redis==4.6.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
retrying==1.3.4
rpds-py==0.25.1
rsa==4.9.1
s3transfer==0.13.0
scikit-learn==1.7.0
scipy==1.15.3
six==1.17.0
slowapi==0.1.9
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.13.1
soxr==0.5.0.post1
SQLAlchemy==2.0.41
starlette==0.46.2
submitit==1.5.3
sympy==1.14.0
tabulate==0.9.0
tenacity==8.5.0
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
toml==0.10.2
toolz==1.0.0
torch==2.7.1
torchaudio==2.7.1
torchlibrosa==0.1.0
tqdm==4.67.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.3
uvloop==0.21.0
vine==5.1.0
watchfiles==1.1.0
wcwidth==0.2.13
websockets==15.0.1
wrapt==1.17.2
